requirements.txt
asgiref==3.6.0
coverage==7.0.3
Django==4.1.5
django-auditlog==2.2.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
